Move darwin environment settings to init-environment-darwin.el and add DICTIONARY=en_US to environment.
I moved the environment settings for darwin from init-environment to a seperate file that is required conditionaly by init-environment. This will allow the system environment settings to be changed and not have an effect on any of the other systems.

I am also now setting the DICTIONARY environment variable to en_US, without this set, hunspell was just hanging and emacs had to be force quit.