Added when clause to conditionaly use package.
I have refactored how I handle packages that are dependant on external
tools being installed. I wrap the use package with a when clause and use
executable-find to test for a primary tool that the package is dependant
on. So for example if python is not found, no reason to use the elpy package.